requirements.txt
numpy==1.17.2
Pillow==6.1.0
